requirements.txt
# Core Dependencies
python-dotenv==1.0.0
groq==0.4.2
huggingface-hub==0.20.3
requests==2.31.0

# UI & Visualization
streamlit==1.30.0
plotly==5.18.0
pandas==2.1.4

# Trend APIs
pytrends==4.9.2
praw==7.7.1

# Image Processing
Pillow==10.2.0

# Utilities
python-dateutil==2.8.2